Move replace item stack method
